[GAIA-2427] Fix RatingDecorator.test.ts
