+++ src/lib/renderer/__tests__/RatingDecorator.test.ts
@@ -6,7 +6,7 @@ import {NoopRating} from "../../renderable/rating/NoopRating";
 import each from "jest-each";
 import {RatingDecorator, RatingRenderStrategy} from "../decorator/RatingDecorator";
 import {ContentCentricRenderer} from "../ContentCentricRenderer";
-import {ISpecification} from "../../api";
+import {RatingTestSpecGenerator} from "./RatingTestSpecGenerator";
 
 
 describe("RatingDecorator test", () => {
@@ -58,6 +58,26 @@ describe("RatingDecorator test", () => {
         expectNotToBeRatingElement(rendered.pop());
     });
 
+    it("do not render rating if enriched is incomplete", () => {
+        const renderer = new RatingDecorator(new ContentCentricRenderer(), RatingRenderStrategy.ALL_EXCEPT_DISABLED_RATINGS);
+        const specification = RatingTestSpecGenerator.generate(true, true);
+        delete specification["enriched"].nodeId;
+
+        const rendered = renderer.render(specification);
+
+        expectNotToBeRatingElement(rendered.pop());
+    });
+
+    it("do not render rating if enriched does not exist", () => {
+        const renderer = new RatingDecorator(new ContentCentricRenderer(), RatingRenderStrategy.ALL_EXCEPT_DISABLED_RATINGS);
+        const specification = RatingTestSpecGenerator.generate(true, true);
+        delete specification["enriched"];
+
+        const rendered = renderer.render(specification);
+
+        expectNotToBeRatingElement(rendered.pop());
+    });
+
     function expectToBeRatingElement(element?: HTMLElement) {
         expect(element).not.toBeNull();
         // @ts-ignore
@@ -77,44 +97,5 @@ describe("RatingDecorator test", () => {
         Renderables.register("container", undefined);
     });
 });
-
-class RatingTestSpecGenerator {
-    static generate(withRatingMarkup: boolean, ratingEnabled: boolean) {
-        const content = [{
-            class: "some-class",
-            type: "block",
-            elements: [
-                {
-                    class: "uppercase bold",
-                    text: "Some very informative text",
-                    type: "label"
-                },
-                {
-                    text: "More very informative text",
-                    type: "label"
-                }
-            ]
-        }];
-
-        if (withRatingMarkup) {
-            return RatingTestSpecGenerator.wrapInContainer([{
-                type: "rating",
-                enabled: ratingEnabled,
-                elements: content
-            }]);
-        }
-
-        return RatingTestSpecGenerator.wrapInContainer(content);
-    }
-
-    private static wrapInContainer(elements: ISpecification[]) {
-        return {
-            type: "container",
-            qualifier: "prompt:next",
-            position: "left",
-            elements
-        } as ISpecification;
-    }
-}
